+++ apps/web/src/pages/api/infinity/pools.ts
@@ -1,6 +1,7 @@
 import BN from 'bignumber.js'
 import { NextRequest, NextResponse } from 'next/server'
 import qs from 'qs'
+import { edgeQueries } from 'quoter/utils/edgePoolQueries'
 
 export const config = {
   runtime: 'edge',
@@ -8,17 +9,6 @@ export const config = {
 
 const MAX_CACHE_SECONDS = 60 * 60
 const BASE_URL = 'https://explorer.pancakeswap.com/api/cached/pools/list'
-
-export type APIChain =
-  | 'bsc'
-  | 'bsc-testnet'
-  | 'ethereum'
-  | 'base'
-  | 'opbnb'
-  | 'zksync'
-  | 'polygon-zkevm'
-  | 'linea'
-  | 'arbitrum'
 
 export default async function handler(req: NextRequest) {
   const raw = new URL(req.url).search.slice(1)
@@ -38,7 +28,7 @@ export default async function handler(req: NextRequest) {
 
   try {
     // eslint-disable-next-line no-await-in-loop
-    const pools = await fetchAllPools({
+    const pools = await edgeQueries.fetchAllPools({
       baseUrl: BASE_URL,
       protocols: protocols as ('infinityBin' | 'infinityCl')[],
       chains: [chain as any],
@@ -64,174 +54,5 @@ export default async function handler(req: NextRequest) {
     )
   } catch (err) {
     return NextResponse.json({ error: `${err}` }, { status: 500, headers: { 'Content-Type': 'application/json' } })
-  }
-}
-
-type PaginatedResponse = {
-  startCursor?: string
-  endCursor?: string
-  hasNextPage: boolean
-  hasPrevPage: boolean
-  rows: Pool[]
-}
-
-type Pool = {
-  id: string
-  chainId: number
-  token0Price: string
-  token1Price: string
-  tvlToken0: string
-  tvlToken1: string
-  tvlUSD: string
-  volumeUSD24h: string
-  apr24h: string
-  protocol: 'v2' | 'v3' | 'infinityBin' | 'infinityCl' | 'stable'
-  feeTier: number
-  token0: Token
-  token1: Token
-  isDynamicFee?: boolean
-  hookAddress?: string | null
-}
-
-type Token = {
-  id: string
-  symbol: string
-  name: string
-  decimals: number
-}
-
-type FetchAllPoolsParams = {
-  baseUrl: string
-  orderBy?: 'tvlUSD' | 'volumeUSD24h' | 'apr24h'
-  protocols: Array<'v2' | 'v3' | 'infinityBin' | 'infinityCl' | 'stable'>
-  chains: Array<
-    'bsc' | 'bsc-testnet' | 'ethereum' | 'base' | 'opbnb' | 'zksync' | 'polygon-zkevm' | 'linea' | 'arbitrum'
-  >
-  pools?: string[]
-  tokens?: string[]
-  pageSize?: number
-  maxPages?: number // Optional safety limit for maximum pages to fetch
-}
-
-/**
- * Fetches all data from a paginated API endpoint
- * @param params Configuration parameters for the fetch operation
- * @returns Promise resolving to an array of all pools
- */
-async function fetchAllPools({
-  baseUrl,
-  orderBy = 'tvlUSD',
-  protocols,
-  chains,
-  pools = [],
-  tokens = [],
-  pageSize = 100,
-  maxPages = Infinity,
-}: FetchAllPoolsParams): Promise<Pool[]> {
-  const allResults: Pool[] = []
-  let cursor: string | null = null
-  let hasNextPage = true
-  let pageCount = 0
-
-  // Construct the base URL params
-  const buildUrlParams = (after?: string) => {
-    const params = new URLSearchParams()
-
-    // Add required parameters
-    params.append('orderBy', orderBy)
-
-    // Add protocols
-    protocols.forEach((protocol) => {
-      params.append('protocols', protocol)
-    })
-
-    // Add chains if tokens are not specified
-    if (tokens.length === 0) {
-      chains.forEach((chain) => {
-        params.append('chains', chain)
-      })
-    }
-
-    // Add pools if specified
-    pools.forEach((pool) => {
-      params.append('pools', pool)
-    })
-
-    // Add tokens if specified
-    tokens.forEach((token) => {
-      params.append('tokens', token)
-    })
-
-    // Add pagination parameters
-    if (after) {
-      params.append('after', after)
-    }
-
-    // Add page size
-    params.append('limit', pageSize.toString())
-
-    return params.toString()
-  }
-
-  while (hasNextPage && pageCount < maxPages) {
-    const url = `${baseUrl}?${buildUrlParams(cursor || undefined)}`
-
-    try {
-      // eslint-disable-next-line no-await-in-loop
-      const response = await fetch(url, {
-        headers: {
-          'x-api-key': process.env.EXPLORER_API_KEY || '',
-          'Content-Type': 'application/json',
-        },
-      })
-
-      if (!response.ok) {
-        throw new Error(`API request failed with status ${response.status}`)
-      }
-
-      // eslint-disable-next-line no-await-in-loop
-      const data: PaginatedResponse = await response.json()
-
-      // Add the current page of results
-      allResults.push(...data.rows)
-
-      // Update for next iteration
-      hasNextPage = data.hasNextPage
-      cursor = data.endCursor || null
-      pageCount++
-    } catch (error) {
-      console.error('Error fetching data:', error)
-      throw error
-    }
-  }
-
-  if (pageCount >= maxPages && hasNextPage) {
-    console.warn(`Reached maximum page limit of ${maxPages}. Some data may not have been fetched.`)
-  }
-
-  return allResults
-}
-
-export const poolTvlMap = async (
-  protocols: ('v2' | 'v3' | 'infinityBin' | 'infinityCl' | 'stable')[],
-  chain: APIChain,
-) => {
-  try {
-    const remotePools = await fetchAllPools({
-      baseUrl: 'https://explorer.pancakeswap.com/api/cached/pools/list',
-      protocols,
-      chains: [chain],
-      orderBy: 'tvlUSD',
-      pageSize: 100,
-    })
-    const tvlMap: Record<string, string> = {}
-    for (const pool of remotePools) {
-      const tvlUSD = pool.tvlUSD
-      const id = pool.id
-      tvlMap[id] = tvlUSD
-    }
-    return tvlMap
-  } catch (ex) {
-    return {}
   }
 }
